boost coverage of csc/cmd to 86.3%

+++ mock/service/controller.go
@@ -23,6 +23,10 @@ func (s *serviceClient) CreateVolume(
 	req *csi.CreateVolumeRequest, _ ...grpc.CallOption) (
 	*csi.CreateVolumeResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock CreateVolume")
+	}
 	return s.service.CreateVolume(ctx, req)
 }
 
@@ -78,6 +82,10 @@ func (s *serviceClient) DeleteVolume(
 	req *csi.DeleteVolumeRequest, _ ...grpc.CallOption) (
 	*csi.DeleteVolumeResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock DeleteVolume")
+	}
 	return s.service.DeleteVolume(ctx, req)
 }
 
@@ -110,6 +118,10 @@ func (s *serviceClient) ControllerPublishVolume(
 	req *csi.ControllerPublishVolumeRequest, _ ...grpc.CallOption) (
 	*csi.ControllerPublishVolumeResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ControllerPublishVolume")
+	}
 	return s.service.ControllerPublishVolume(ctx, req)
 }
 
@@ -157,6 +169,10 @@ func (s *serviceClient) ControllerUnpublishVolume(
 	req *csi.ControllerUnpublishVolumeRequest, _ ...grpc.CallOption) (
 	*csi.ControllerUnpublishVolumeResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ControllerUnpublishVolume")
+	}
 	return s.service.ControllerUnpublishVolume(ctx, req)
 }
 
@@ -206,6 +222,10 @@ func (s *serviceClient) ValidateVolumeCapabilities(
 	req *csi.ValidateVolumeCapabilitiesRequest, _ ...grpc.CallOption) (
 	*csi.ValidateVolumeCapabilitiesResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ValidateVolumeCapabilities")
+	}
 	return s.service.ValidateVolumeCapabilities(ctx, req)
 }
 
@@ -228,6 +248,10 @@ func (s *serviceClient) ListVolumes(
 	req *csi.ListVolumesRequest, _ ...grpc.CallOption) (
 	*csi.ListVolumesResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ListVolumes")
+	}
 	return s.service.ListVolumes(ctx, req)
 }
 
@@ -312,6 +336,10 @@ func (s *serviceClient) GetCapacity(
 	req *csi.GetCapacityRequest, _ ...grpc.CallOption) (
 	*csi.GetCapacityResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock GetCapacity")
+	}
 	return s.service.GetCapacity(ctx, req)
 }
 
@@ -330,6 +358,10 @@ func (s *serviceClient) ControllerGetCapabilities(
 	req *csi.ControllerGetCapabilitiesRequest, _ ...grpc.CallOption) (
 	*csi.ControllerGetCapabilitiesResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ControllerGetCapabilities")
+	}
 	return s.service.ControllerGetCapabilities(ctx, req)
 }
 
@@ -391,6 +423,10 @@ func (s *serviceClient) CreateSnapshot(
 	req *csi.CreateSnapshotRequest, _ ...grpc.CallOption) (
 	*csi.CreateSnapshotResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock CreateSnapshot")
+	}
 	return s.service.CreateSnapshot(ctx, req)
 }
 
@@ -414,6 +450,10 @@ func (s *serviceClient) DeleteSnapshot(
 	req *csi.DeleteSnapshotRequest, _ ...grpc.CallOption) (
 	*csi.DeleteSnapshotResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock DeleteSnapshot")
+	}
 	return s.service.DeleteSnapshot(ctx, req)
 }
 
@@ -434,6 +474,10 @@ func (s *serviceClient) ListSnapshots(
 	req *csi.ListSnapshotsRequest, _ ...grpc.CallOption) (
 	*csi.ListSnapshotsResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ListSnapshots")
+	}
 	return s.service.ListSnapshots(ctx, req)
 }
 
@@ -521,6 +565,10 @@ func (s *serviceClient) ControllerExpandVolume(
 	req *csi.ControllerExpandVolumeRequest, _ ...grpc.CallOption) (
 	*csi.ControllerExpandVolumeResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ControllerExpandVolume")
+	}
 	return s.service.ControllerExpandVolume(ctx, req)
 }
 
@@ -565,6 +613,10 @@ func (s *serviceClient) ControllerGetVolume(
 	req *csi.ControllerGetVolumeRequest, _ ...grpc.CallOption) (
 	*csi.ControllerGetVolumeResponse, error,
 ) {
+	// if CTX has this key, we want to return error for UT
+	if ctx.Value(ContextKey("returnError")) == "true" {
+		return nil, status.Error(codes.InvalidArgument, "Returned error from mock ControllerGetVolume")
+	}
 	return s.service.ControllerGetVolume(ctx, req)
 }
